add initial working toggle
