+++ ampere/viz.py
@@ -347,110 +347,6 @@ def viz_follower_network(show_fig: bool = False) -> Figure:
     return fig
 
 
-def viz_summary(
-    df: pd.DataFrame,
-    metric_type: str,
-    date_range: Optional[list[int]] = None,
-    show_fig: bool = False,
-    screen_width: ScreenWidth = ScreenWidth.lg,
-) -> Figure:
-    df_filtered = df.query(f"metric_type == '{metric_type}'").sort_values("metric_date")
-    if date_range is not None:
-        filter_date_min = datetime.datetime.fromtimestamp(
-            date_range[0], tz=pytz.timezone("America/New_York")
-        )
-        filter_date_max = datetime.datetime.fromtimestamp(
-            date_range[1], tz=pytz.timezone("America/New_York")
-        )
-
-        df_filtered = df_filtered.query(f"metric_date >= '{filter_date_min}'").query(
-            f"metric_date <= '{filter_date_max}'"
-        )
-
-    repo_palette = generate_repo_palette()
-    fig = px.area(
-        df_filtered,
-        x="metric_date",
-        y="metric_count",
-        color="repo_name",
-        template="simple_white",
-        hover_name="repo_name",
-        color_discrete_map=repo_palette,
-        height=500,
-        category_orders={"repo_name": repo_palette.keys()},
-        facet_col="metric_type" # single var facet col for plot title
-    )
-
-    fig.for_each_annotation(
-        lambda a: a.update(
-            text="<b>"
-            + a.text.split("=")[-1]
-            + "</b>",
-            font_size=18,
-            bgcolor=AmperePalette.PAGE_ACCENT_COLOR2,
-            font_color="white",
-            borderpad=5,
-        )
-    )
-    fig.update_yaxes(matches=None, showticklabels=True)
-    fig.update_traces(hovertemplate="<b>%{x}</b><br>n=%{y}")
-
-    fig_legend_y = {ScreenWidth.xs: 1.04, ScreenWidth.sm: 1.02}
-    if screen_width in [ScreenWidth.xs, ScreenWidth.sm]:
-        fig.update_layout(
-            legend=dict(
-                title=None,
-                itemsizing="constant",
-                font=dict(size=14),
-                orientation="h",
-                yanchor="top",
-                y=fig_legend_y[screen_width],
-                xanchor="center",
-                x=0.5,
-            )
-        )
-    else:
-        fig.update_layout(
-            legend=dict(title=None, itemsizing="constant", font=dict(size=14))
-        )
-
-    fig.for_each_annotation(
-        lambda a: a.update(
-            text="<b>" + a.text.split("=")[-1] + "</b>",
-            font_size=18,
-            bgcolor=AmperePalette.PAGE_ACCENT_COLOR2,
-            font_color="white",
-            borderpad=5,
-        )
-    )
-
-    fig.for_each_yaxis(
-        lambda y: y.update(
-            title="",
-            showline=True,
-            linewidth=1,
-            linecolor="black",
-            mirror=True,
-            tickfont_size=14,
-        )
-    )
-    fig.for_each_xaxis(
-        lambda x: x.update(
-            title="",
-            showline=True,
-            linewidth=1,
-            linecolor="black",
-            mirror=True,
-            showticklabels=True,
-            tickfont_size=14,
-        )
-    )
-
-    fig.update_layout(margin=dict(l=0, r=0))
-    if show_fig:
-        fig.show()
-
-    return fig
 
 
 NETWORK_LAYOUT = go.Layout(
